requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
CacheControl==0.14.4
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==3.0.5
channels-redis>=4.2.0
chardet==5.2.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==3.0.2
dj-config-url==0.1.1
dj-database-url==3.1.2
Django==5.0.2
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
firebase_admin==7.1.0
frozenlist==1.8.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==25.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
multidict==6.7.1
mysqlclient==2.2.8
packaging==26.0
pillow==12.1.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.9
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dotenv==1.2.2
pytz==2025.2
razorpay==2.0.0
reportlab==4.1.0
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
setuptools==68.2.2
sqlparse==0.5.5
twilio==8.1.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
wheel==0.46.3
whitenoise==6.11.0
yarl==1.22.0
zope.interface==8.2
PyMySQL==1.1.2